refactor: PIZ5.1 Changed utils functions to faker functions
