+++ app/test/services/test_beverage.py
@@ -1,6 +1,7 @@
 import pytest
+from faker import Faker
 
-from app.test.utils.functions import get_random_string, get_random_price
+fake = Faker()
 
 
 def test_create_beverage_service(create_beverage):
@@ -13,7 +14,8 @@ def test_create_beverage_service(create_beverage):
 
 def test_update_beverage_service(client, create_beverage, beverage_uri):
     current_beverage = create_beverage.json
-    update_data = {**current_beverage, 'name': get_random_string(), 'price': get_random_price(1, 5)}
+    update_data = {**current_beverage, 'name': fake.pystr(),
+                   'price': fake.pyfloat(left_digits=2, right_digits=3, positive=True)}
     response = client.put(beverage_uri, json=update_data)
     pytest.assume(response.status.startswith('200'))
     updated_beverage = response.json
