+++ app/test/services/test_size.py
@@ -1,6 +1,9 @@
 import pytest
+from faker import Faker
 
-from app.test.utils.functions import get_random_string, get_random_price
+fake = Faker()
+
+
 
 
 def test_create_size_service(create_size):
@@ -13,8 +16,8 @@ def test_create_size_service(create_size):
 
 def test_update_size_service(client, create_size, size_uri):
     current_size = create_size.json
-    update_data = {**current_size, 'name': get_random_string(),
-                   'price': get_random_price(1, 5)}
+    update_data = {**current_size, 'name': fake.pystr(),
+                   'price': fake.pyfloat(left_digits=2, right_digits=3, positive=True)}
     response = client.put(size_uri, json=update_data)
     pytest.assume(response.status.startswith('200'))
     updated_size = response.json
