+++ requirements.txt
@@ -3,6 +3,7 @@ attrs==21.4.0
 autopep8==1.6.0
 click==8.0.3
 coverage==6.6.0b1
+Faker==15.2.0
 flake8==4.0.1
 Flask==2.0.2
 Flask-Cors==3.0.10
